+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Fix: Log real exitCode, stdout and stdout if available (#13)
+
 # 1.0.0-alpha.3
 
 * Bump sentry-cli 1.69.1 which includes a fix for Dart debug symbols (#8)